requirements.txt
numpy==1.13
scipy==0.17.1
pymorphy2[fast]
nltk
gensim
tensorflow
h5py
sklearn-pandas
pandas==0.21
pyaml
keras
future
beautifulsoup4
alphabet-detector
webhoseio
